% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_connect_simple.R
\name{assert_db_connected}
\alias{assert_db_connected}
\title{assert_db_connected}
\usage{
assert_db_connected(con = "psql_con")
}
\arguments{
\item{con}{string of global variable containing the connection, defaults to psql_con created by connect_database}
}
\description{
assert_db_connected
}
