% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_plotting.R
\name{blur_color}
\alias{blur_color}
\title{This function will generate a color close to the original color provided}
\usage{
blur_color(color)
}
\arguments{
\item{color}{color to randomly adjust (as hex color)}
}
\description{
This function will generate a color close to the original color provided
}
