% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{clean_refseq_genus}
\alias{clean_refseq_genus}
\title{Will generate a new taxonomic table where the genus names are tidied up
interpretable names (prefaced by the taxonomic order thy come from).
This function should be used for cases where the taxonomy is based on
refseq blast hits, such as our amplicon pipeline,}
\usage{
clean_refseq_genus(phy_seq_obj)
}
\arguments{
\item{phy_seq_obj}{A phyloseq object with taxonomic and otu information}
}
\description{
Will generate a new taxonomic table where the genus names are tidied up
interpretable names (prefaced by the taxonomic order thy come from).
This function should be used for cases where the taxonomy is based on
refseq blast hits, such as our amplicon pipeline,
}
