% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{clean_SGB_genus}
\alias{clean_SGB_genus}
\title{Will generate a new taxonomic table where the SGB mice names have been replaced with more
interpretable names (prefaced by the taxonomic order they come from).}
\usage{
clean_SGB_genus(phy_seq_obj)
}
\arguments{
\item{phy_seq_obj}{A phyloseq object with taxonomic and otu information}
}
\description{
Will generate a new taxonomic table where the SGB mice names have been replaced with more
interpretable names (prefaced by the taxonomic order they come from).
}
