% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_beta_diversity_and_tsne.R
\name{compute_beta_diversity_and_tsne}
\alias{compute_beta_diversity_and_tsne}
\title{This function helps you to compute beta-diversity, tsne and PCA plots}
\usage{
compute_beta_diversity_and_tsne(sampleid, taxonomy, count)
}
\arguments{
\item{sampleid}{vector of sampleid. sampleid, taxonomy and cout must have same length}

\item{taxonomy}{taxonomic classification. sampleid, taxonomy and cout must have same length}

\item{count}{count. sampleid, taxonomy and cout must have same length}
}
\description{
This function helps you to compute beta-diversity, tsne and PCA plots
}
\examples{
# We show how to compute a tsne map with 100 samples from
# frozen_set_ag based on the beta-diversity at genus level
# getting samples from frozen set
connect_database(bundled=TRUE)
get_table_from_database("frozen_set_ag") 
asv_counts_subset = get_counts_subset( frozen_set_ag$sampleid[1:100])
get_table_from_database("asv_annotation_blast_ag"); #getting annotation
m = merge(asv_counts_subset[,.(asv_key,sampleid,count)],asv_annotation_blast_ag[,.(asv_key,genus)])
cbd = compute_beta_diversity_and_tsne(m$sampleid, m$genus, m$count)
cbd$compute_beta_diversity();
cbd$compute_tsne(perplexity = 2, max_iter = 500)
data_set_tsne = cbd$get_tsne();

}
