% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_connect_simple.R
\name{connect_database}
\alias{connect_database}
\title{connect_database}
\usage{
connect_database(config_file, bundled = FALSE)
}
\arguments{
\item{config_file}{file with login info to connect to database}

\item{bundled}{conected to test sqlite database}
}
\description{
connect_database
}
