% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_abx.R
\name{get_abx_from_tasks}
\alias{get_abx_from_tasks}
\title{get_abx_from_tasks}
\usage{
get_abx_from_tasks(tasks_path)
}
\arguments{
\item{tasks_path}{path to tasks pull in rds format, such as /deep_sequencing/Clinical Annotation/Abx_data_TsoniIDB_pull_2020_07-06/allo_tasks_2020-07-10.rds}
}
\description{
get_abx_from_tasks
}
