% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_isabl_analyses}
\alias{get_isabl_analyses}
\title{Get all analyses for a set of samples, optionally filering by application}
\usage{
get_isabl_analyses(sampleids, app_id = NA, proj_id = NA)
}
\arguments{
\item{sampleids}{vector of sample identifiers}

\item{app_id}{isabl app identifier}

\item{proj_id}{isabl project to filter by}
}
\description{
Get all analyses for a set of samples, optionally filering by application
}
\examples{
\dontrun{
connect_database()
sampleids <- c("2711D", "2711E", "2711F", "2711G", "2711H", "2711I", "2954A", 
              "2954B", "2954C", "2954D", "3384D", "3384E", "3384F", "3384H")
 get_isabl_analyses(sampleids = sampleids, app_id=43)
 }
}
