% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_project_samples}
\alias{get_project_samples}
\title{For a given project number, get all the samples assocoiated with it - optionally
filtering by those samples which have successful analyses run through a particular
application number.}
\usage{
get_project_samples(project, app_id = NA, verbose = FALSE)
}
\arguments{
\item{project}{the project number to fetch analysis ids for}

\item{app_id}{the number of the application to check samples have a successful analysis for.}

\item{verbose}{if messages should be printed about what is being run/found.}
}
\description{
For a given project number, get all the samples assocoiated with it - optionally
filtering by those samples which have successful analyses run through a particular
application number.
}
\examples{
# ie to get all the Human Biobakery results for project 32: 
\dontrun{
ia_ids = get_project_samples(project = 32, app_id = 66, verbose = T)
}
}
