% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_sample_isabl_info}
\alias{get_sample_isabl_info}
\title{Given sample-ids - get the associated experiments and analyses associated with those experiments (optionally filtering by project or app_id)}
\usage{
get_sample_isabl_info(
  sample_ids,
  verbose = FALSE,
  app_id = NA,
  proj_id = NA,
  schema = "public",
  allow_excluded = FALSE
)
}
\arguments{
\item{sample_ids}{list of sample ids}

\item{verbose}{output additional logging information (default = F)}

\item{app_id}{isabl app identifier (optional)}

\item{proj_id}{isabl project to filter by}

\item{schema}{name of schema to query; public for production tables, test_tables for dev, and main when using sqlitedb}

\item{allow_excluded}{Attempt to include analyses tagged for exclusion}
}
\description{
Given sample-ids - get the associated experiments and analyses associated with those experiments (optionally filtering by project or app_id)
}
\examples{
# assuming test_metadata is a dataframe with a column containing sample IDs called samplid
\dontrun{
get_sample_isabl_info(list(db[, sample_id_col]), app_id = 66)
}
}
