% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_subset_pg_df}
\alias{get_subset_pg_df}
\title{Helper function that speeds up interactions with the postgres database by only
fetching those parts of the table which contain a list of desired IDs.}
\usage{
get_subset_pg_df(table_name, table_id_name, ids, con = NA, schema = "public")
}
\arguments{
\item{table_name}{the name of the table to do the subsetting on.}

\item{table_id_name}{the name of the id in the table to perform filtering with.}

\item{ids}{list of strings: ids to use in the query}

\item{con}{\itemize{
\item the postgres connection object, defaults to psql_con
}}

\item{schema}{name of schema to query; public for production tables, test_tables for dev}
}
\description{
Helper function that speeds up interactions with the postgres database by only
fetching those parts of the table which contain a list of desired IDs.
}
