% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_table_from_database.R
\name{get_table_from_database}
\alias{get_table_from_database}
\alias{list_table_from_database}
\alias{get_counts_subset}
\title{Querying database}
\usage{
get_table_from_database(table_name, pre_filter = TRUE)

list_table_from_database(pattern = NULL)

get_counts_subset(sampleid_subset = NULL, pre_filter = T)
}
\arguments{
\item{table_name}{name of table to be loaded.}

\item{pre_filter}{If True for samples_castori_ag, it filters out sampleids that appear in samples_castori_blacklist.}

\item{pattern}{only tables that match \emph{pattern} will be returned.}

\item{sampleid_subset}{Only sampleids in this vector will be returned.}
}
\description{
Querying database
}
