% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_beta_diversity_and_tsne.R
\name{make_tsne}
\alias{make_tsne}
\title{This function helps you to compute beta-diversity, tsne and PCA plots}
\usage{
make_tsne(dists, perplexity = 20, max_iter = 3000, theta = 0.1, seed = 1)
}
\arguments{
\item{dists}{distance matrix from the phyloseq/vegan::distance function}

\item{perplexity}{rtsne's perplexity argument}

\item{max_iter}{rtsne's max_iter argument}

\item{theta}{rtsne's theta argument}

\item{seed}{rtsne's seed argument}
}
\description{
This function helps you to compute beta-diversity, tsne and PCA plots
}
