% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_plotting.R
\name{norm_generate_component}
\alias{norm_generate_component}
\title{This function will generate a color close to the original color comp provided}
\usage{
norm_generate_component(rgb_component, magnitude = 15)
}
\arguments{
\item{rgb_component}{the component to randomly adjust}

\item{magnitude}{the maximum amount to adjust rgb component by (default 15)}
}
\description{
This function will generate a color close to the original color comp provided
}
