% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{padMRN}
\alias{padMRN}
\title{Pad MRNs with zeros to avoid the conflicts between dbs storing mrns as characters vs as integers.}
\usage{
padMRN(mrn)
}
\arguments{
\item{mrn}{string or integer of identifier}
}
\description{
Pad MRNs with zeros to avoid the conflicts between dbs storing mrns as characters vs as integers.
}
\examples{
mrns <- c(1, "42")
padMRN(mrns)
#  "00000001" "00000042"

}
