% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{process_metadata}
\alias{process_metadata}
\title{Ensure that metadata is a dataframe without NA sampleid or duplicated sampleids.
cast metadata to dataframe so we can do convenient indexing}
\usage{
process_metadata(metadata, sampleid_col)
}
\arguments{
\item{metadata}{metadata dataframe}

\item{sampleid_col}{name of column in metadata containing sample_ids}
}
\description{
Ensure that metadata is a dataframe without NA sampleid or duplicated sampleids.
cast metadata to dataframe so we can do convenient indexing
}
