% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{vdb_make_phylo_mgx}
\alias{vdb_make_phylo_mgx}
\title{Make a phyloseq object from some metadata with sampleids}
\usage{
vdb_make_phylo_mgx(
  metadata,
  sampleid_col = "sampleid",
  app_id = 66,
  verbose = FALSE,
  choose_max_experiment = FALSE,
  testing = FALSE
)
}
\arguments{
\item{metadata}{dataframe with at least one column sampleids; everything else is treated as sample info}

\item{sampleid_col}{name of the column containing sampleids}

\item{app_id}{isabl app identifier}

\item{verbose}{output additional logging information}

\item{choose_max_experiment}{if multiple analyses are found, select the one utilixing the most experiemnts}

\item{testing}{use test tables if TRUE}
}
\description{
Make a phyloseq object from some metadata with sampleids
}
\examples{
# assuming test_metadata is a dataframe with a column containing sample IDs called samplid
\dontrun{
vdb_make_phylo_mgx(test_metadata, sampleid_col = "sampleid")
}
}
