% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vdbR-package.R
\docType{package}
\name{vdbR-package}
\alias{vdbR}
\alias{vdbR-package}
\title{vdbR: A Package for Acessing and Processing Microbiome Data}
\description{
Utility functions for dealing with microbiome data.
}
\author{
\strong{Maintainer}: Antonio Gomes \email{gomesa@mskcc.org} (\href{https://orcid.org/0000-0003-3790-3724}{ORCID})

Authors:
\itemize{
  \item Anqi Dai \email{daia1@mskcc.org} (\href{https://orcid.org/0000-0003-3106-4624}{ORCID})
  \item Chi Nguyen \email{nguyenc1@sloankettering.edu}
  \item Anastasia Kousa \email{kousaa@mskcc.org} (\href{https://orcid.org/0000-0002-4550-0389}{ORCID})
  \item Nick Waters \email{watersn@mskcc.org} (\href{https://orcid.org/0000-0002-9035-2143}{ORCID})
  \item Mirae Baichoo \email{baichom1@mskcc.org} (\href{https://orcid.org/0000-0002-4219-9076}{ORCID})
}

}
\keyword{internal}
