% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{full_palette}
\alias{full_palette}
\title{Default palette}
\format{
\subsection{\code{full_palette}}{
}
}
\source{
"antonio gomes (color tables and code)"
}
\usage{
full_palette
}
\description{
Default palette for vdbR contains previously published colors along with default palette from Ying Taur
}
\keyword{datasets}
