% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_plotting.R
\name{rename_taxa_colors}
\alias{rename_taxa_colors}
\title{This function will generate colors for the taxa_levels provided based on the base_colors and the taxa levels.
New colors will be generated in a spread on the appropriate rank.}
\usage{
rename_taxa_colors(palette, full_taxonomy, rank)
}
\arguments{
\item{palette}{the palette to overwrite colors for.}

\item{full_taxonomy}{full taxonomy rows for the species to generate colors for.}

\item{rank}{the rank of the palette}
}
\description{
This function will generate colors for the taxa_levels provided based on the base_colors and the taxa levels.
New colors will be generated in a spread on the appropriate rank.
}
